+++ test_settings.py
@@ -29,9 +29,7 @@ if django.VERSION >= (1, 8):
                     'django.template.context_processors.debug',
                     'django.template.context_processors.request',
                     'django.contrib.auth.context_processors.auth',
-                    'django.contrib.messages.context_processors.messages',
-                    'allauth.account.context_processors.account',
-                    'allauth.socialaccount.context_processors.socialaccount',
+                    'django.contrib.messages.context_processors.messages'
                 ],
             },
         },
@@ -45,9 +43,6 @@ else:
         "django.core.context_processors.static",
         "django.core.context_processors.request",
         "django.contrib.messages.context_processors.messages",
-    
-        "allauth.account.context_processors.account",
-        "allauth.socialaccount.context_processors.socialaccount",
     )
 
 MIDDLEWARE_CLASSES = (